refactor: extract ProxyClient into AppState extension (Stage 3c)
Stage 3c — last of the Stage 3 sub-extractions. Moves all proxy HTTP +
NIP-98 + retry-queue logic out of AppState.swift into a dedicated
extension file. The new file owns the [Pair] log category from PR #27.

Moves out of Clave/AppState.swift:
- registerWithProxy(completion:) — single-account wrapper
- registerAllAccountsWithProxy() — multi-account fan-out
- registerSignerWithProxy(signer:completion:) — private NIP-98 POST
  /register impl
- ensureRegisteredFresh() — throttled scenePhase trigger
- ensureAllRegisteredFresh() — multi-account variant
- unregisterWithProxy(signer:) — NIP-98 POST /unregister
- pairClientWithProxy(clientPubkey:relayUrls:signer:) — POST /pair-client
  with retry queue on failure
- unpairClientWithProxy(clientPubkey:signer:) — POST /unpair-client
- drainPendingPairOps() — retry queue worker
- retryPairOp(op:relayUrls:) — private retry impl
- retryUnpairOp(op:) — private retry impl
- "// MARK: - Proxy per-client-relay (V2)" header
- private let logger = Logger(...category: "pair") relocated from
  AppState.swift line 8 (16/16 of its uses moved with the cluster)

Into the new file:
- Clave/AppState+ProxyClient.swift (577 LOC) — extension AppState

Logger relocation note: declaring Logger(subsystem:..., category:...)
in two files is equivalent — multiple Logger instances with the same
subsystem + category combine into one log stream. The dev-menu "Copy
Recent Logs" path (LogExporter.allCategories includes "pair") is
unchanged.

Zero behavior change:
- Function bodies preserved byte-for-byte.
- Original `private` modifiers preserved on registerSignerWithProxy,
  retryPairOp, retryUnpairOp.
- All call sites (network code, Task { @MainActor in } closures,
  URLSession callbacks, NIP-98 signing) unchanged.
- Cross-extension call from AppState+NostrConnect.swift's
  handleNostrConnect to pairClientWithProxy still works — both are
  extension AppState in the same module.

External callers unchanged:
- Clave/ClaveApp.swift (drainPendingPairOps via NotificationCenter)
- Clave/Views/MainTabView.swift (ensureRegisteredFresh +
  ensureAllRegisteredFresh on scenePhase)
- Clave/Views/Settings/SettingsView.swift (registerWithProxy,
  unregisterWithProxy)
- Clave/Views/Onboarding/OnboardingView.swift (registerWithProxy)
- Clave/Views/Home/HomeView.swift (registerAllAccountsWithProxy)
- Clave/Views/Home/ClientDetailView.swift (unpairClientWithProxy)
All call via appState.<method> — extension dispatch.

No pbxproj edits — Clave/ directory is auto-synced.

Verification:
- xcodebuild test -skip-testing:ClaveUITests on iPhone 17 / iOS 26.4:
  - Pre-baseline (main @ 11c7117): 232 passed / 0 failed
  - Post-extraction (this branch): 232 passed / 0 failed
  - ** TEST SUCCEEDED ** in both runs

AppState.swift: 1,483 -> 922 LOC (-561, -37.8%).
Combined Stages 1+2+3a+3b+3c: 2,338 -> 922 LOC (-60.6% from original).

Stage 4 (AccountManager + PendingApprovalCoordinator, ~700 LOC) is
the final and highest-risk extraction — touches multi-account
@Observable surface and the alert-chain state machine.
